refactor(core): move four self-contained blocks out of core/common.py
Pure relocation. Every moved name is re-exported from core.common, so no
caller changes.
