+++ geoparquet_io/core/common.py
@@ -1,13 +1,44 @@
+"""Shared helpers for reading and writing GeoParquet.
+
+This module is being taken apart. It grew to 4,138 lines with 50 importers, and
+the deferred (function-level) imports written across ``core/`` to dodge cycles
+through it are the symptom -- a module this central is one every new helper has
+to route around. Blocks that answer a single question and carry no coupling to
+the rest have moved out to modules named for that question.
+
+Every name that moved is still importable from here, so that no caller had to
+change in the commit that moved the code. Those re-exports are written
+``import X as X`` -- the explicit re-export form -- which is both how a reader
+tells a compatibility shim from a helper this module actually uses, and how
+ruff and vulture tell the same thing. The shims are deleted, and their callers
+repointed at the real home, in a later pass. What went where:
+
+* :mod:`geoparquet_io.core.arrow_types` -- Arrow type promotion and casting
+* :mod:`geoparquet_io.core.bbox_structure` -- bbox column / covering detection
+* :mod:`geoparquet_io.core.compression` -- compression option validation
+* :mod:`geoparquet_io.core.file_type` -- GeoParquet file-type detection
+* :mod:`geoparquet_io.core.format_writers` -- ``create_shapefile_zip``
+* :mod:`geoparquet_io.core.sizing` -- byte-size parsing and formatting
+"""
+
 import copy
 import json
 import os
 import re
 from collections.abc import Collection
-from pathlib import Path
-from typing import Any, Literal, TypedDict
+from typing import Any
 
 import duckdb
 import pyarrow.parquet as pq
+
+from geoparquet_io.core.arrow_types import _cast_table_to_schema as _cast_table_to_schema
+from geoparquet_io.core.arrow_types import _compute_unified_schema as _compute_unified_schema
+from geoparquet_io.core.arrow_types import _promote_numeric_type as _promote_numeric_type
+from geoparquet_io.core.arrow_types import _rebuild_array_with_type
+from geoparquet_io.core.bbox_structure import BboxInfo as BboxInfo
+from geoparquet_io.core.bbox_structure import _bbox_column_from_covering, check_bbox_structure
+from geoparquet_io.core.bbox_structure import get_bbox_advice as get_bbox_advice
+from geoparquet_io.core.compression import validate_compression_settings
 
 # Internal imports - used by functions in this module
 from geoparquet_io.core.crs_utils import (
@@ -30,16 +61,19 @@ from geoparquet_io.core.duckdb_utils import (
     validate_compression_level,
 )
 from geoparquet_io.core.exceptions import (
-    FileNotFoundGeoParquetError,
     GeoParquetError,
     InvalidParameterError,
 )
+from geoparquet_io.core.file_type import detect_geoparquet_file_type
+from geoparquet_io.core.file_type import (
+    detect_geoparquet_file_type_cache_clear as detect_geoparquet_file_type_cache_clear,
+)
 from geoparquet_io.core.file_utils import (
-    _get_file_cache_key,
     get_first_parquet_file,
     is_partition_path,
     resolve_file_url,
 )
+from geoparquet_io.core.format_writers import create_shapefile_zip as create_shapefile_zip
 from geoparquet_io.core.geo_metadata import (
     DEFAULT_GEOPARQUET_VERSION,
     GEOPARQUET_VERSIONS,
@@ -86,6 +120,8 @@ from geoparquet_io.core.remote import (
     remote_write_context,
     upload_if_remote,
 )
+from geoparquet_io.core.sizing import format_size as format_size
+from geoparquet_io.core.sizing import parse_size_string as parse_size_string
 from geoparquet_io.core.streaming import extract_version_from_metadata
 
 
@@ -102,139 +138,6 @@ def should_skip_bbox(geoparquet_version):
         bool: True if bbox should be skipped, False if bbox should be added
     """
     return geoparquet_version in ("2.0", "parquet-geo-only", "1.1-geoarrow")
-
-
-# LRU cache for detect_geoparquet_file_type results
-# Using a simple dict cache with manual mtime tracking for invalidation
-_file_type_cache: dict[str, tuple[float, dict]] = {}
-_FILE_TYPE_CACHE_MAX_SIZE = 100
-
-
-def _check_file_type_cache(parquet_file: str) -> dict | None:
-    """Check cache for file type detection result."""
-    cache_key, mtime = _get_file_cache_key(parquet_file)
-    if cache_key in _file_type_cache:
-        cached_mtime, result = _file_type_cache[cache_key]
-        # For remote files (mtime=0), always use cache
-        # For local files, invalidate if file changed
-        if mtime == 0 or cached_mtime == mtime:
-            return result
-    return None
-
-
-def _update_file_type_cache(parquet_file: str, result: dict) -> None:
-    """Update cache with file type detection result."""
-    global _file_type_cache
-    cache_key, mtime = _get_file_cache_key(parquet_file)
-
-    # Simple LRU: if cache is full, clear half of it
-    if len(_file_type_cache) >= _FILE_TYPE_CACHE_MAX_SIZE:
-        # Remove oldest half
-        keys_to_remove = list(_file_type_cache.keys())[: _FILE_TYPE_CACHE_MAX_SIZE // 2]
-        for k in keys_to_remove:
-            del _file_type_cache[k]
-
-    _file_type_cache[cache_key] = (mtime, result)
-
-
-def detect_geoparquet_file_type(parquet_file, verbose=False, con=None):
-    """
-    Detect the GeoParquet/Parquet-geo type of a file.
-
-    Determines whether a file is:
-    - GeoParquet 1.x (has geo metadata with version 1.x)
-    - GeoParquet 2.0 (has geo metadata with version 2.x, uses native Parquet geo types)
-    - Parquet-geo-only (has native Parquet geo types but NO geo metadata)
-    - Unknown (no geo indicators found)
-
-    Performance notes (Issue #232):
-    - For local files, uses PyArrow for ~300x faster metadata reads
-    - Results are cached with mtime-based invalidation
-    - Pass `con` to reuse an existing DuckDB connection for remote files
-
-    Args:
-        parquet_file: Path to the parquet file
-        verbose: Whether to print verbose output
-        con: Optional DuckDB connection to reuse (for remote file operations)
-
-    Returns:
-        dict with:
-            - has_geo_metadata: bool - Has 'geo' key in metadata
-            - geo_version: str - GeoParquet version from metadata (e.g., "1.1.0", "2.0.0") or None
-            - has_native_geo_types: bool - Has Parquet GEOMETRY/GEOGRAPHY logical types
-            - file_type: str - One of: "geoparquet_v1", "geoparquet_v2", "parquet_geo_only", "unknown"
-            - bbox_recommended: bool - Whether bbox column is recommended for this file type
-    """
-    from geoparquet_io.core.duckdb_metadata import (
-        detect_geometry_columns,
-        get_geo_metadata,
-    )
-
-    # Check cache first (skip if connection provided - caller wants fresh read)
-    if con is None:
-        cached_result = _check_file_type_cache(parquet_file)
-        if cached_result is not None:
-            if verbose:
-                debug(f"File type detection (cached): {cached_result}")
-            return cached_result.copy()  # Return copy to prevent mutation
-
-    result = {
-        "has_geo_metadata": False,
-        "geo_version": None,
-        "has_native_geo_types": False,
-        "file_type": "unknown",
-        "bbox_recommended": True,  # Default for v1.x
-    }
-
-    # Check for geo metadata (uses PyArrow for local files, DuckDB for remote)
-    geo_meta = get_geo_metadata(parquet_file, con=con)
-    if geo_meta:
-        result["has_geo_metadata"] = True
-        if isinstance(geo_meta, dict) and "version" in geo_meta:
-            # A version that is not a string is no version this can report:
-            # `2` reached `.startswith` below and took `check spec`,
-            # `check bbox`, `check optimization` and `add bbox` down with a bare
-            # `AttributeError` (#979). `geo_version` is documented as the
-            # version *or None*, and every consumer already handles None.
-            result["geo_version"] = carried_version(geo_meta["version"], source=str(parquet_file))
-
-    # Check for native Parquet geo types using schema
-    geo_columns = detect_geometry_columns(parquet_file, con=con)
-    if geo_columns:
-        result["has_native_geo_types"] = True
-
-    # Determine file type
-    if result["has_geo_metadata"]:
-        version = result["geo_version"]
-        if version and version.startswith("2."):
-            result["file_type"] = "geoparquet_v2"
-            result["bbox_recommended"] = False  # V2 uses native geo row group stats
-        else:
-            result["file_type"] = "geoparquet_v1"
-            result["bbox_recommended"] = True  # V1.x needs bbox for spatial filtering
-    elif result["has_native_geo_types"]:
-        result["file_type"] = "parquet_geo_only"
-        result["bbox_recommended"] = False  # Native geo types provide row group stats
-    # else: remains "unknown"
-
-    # Cache the result (only if no connection provided)
-    if con is None:
-        _update_file_type_cache(parquet_file, result)
-
-    if verbose:
-        debug(f"File type detection: {result}")
-
-    return result
-
-
-def detect_geoparquet_file_type_cache_clear():
-    """Clear the file type detection cache."""
-    global _file_type_cache
-    _file_type_cache = {}
-
-
-# Add cache_clear method to the function for compatibility
-detect_geoparquet_file_type.cache_clear = detect_geoparquet_file_type_cache_clear
 
 
 def get_parquet_metadata(parquet_file, verbose=False):
@@ -333,106 +236,6 @@ def calculate_file_bounds(file_path, geom_column=None, verbose=False):
         return None
     finally:
         con.close()
-
-
-def parse_size_string(size_str):
-    """
-    Parse a human-readable size string into bytes.
-
-    Args:
-        size_str: String like '256MB', '1GB', '128' (assumed MB if no unit)
-
-    Returns:
-        int: Size in bytes
-    """
-    if not size_str:
-        return None
-
-    # Handle plain numbers (assume MB)
-    try:
-        return int(size_str) * 1024 * 1024
-    except ValueError:
-        pass
-
-    # Parse with units
-    size_str = size_str.strip().upper()
-    match = re.match(r"^(\d+(?:\.\d+)?)\s*([KMGT]?B?)$", size_str)
-    if not match:
-        raise ValueError(f"Invalid size format: {size_str}")
-
-    value = float(match.group(1))
-    unit = match.group(2)
-
-    # Convert to bytes
-    multipliers = {
-        "B": 1,
-        "KB": 1024,
-        "MB": 1024 * 1024,
-        "GB": 1024 * 1024 * 1024,
-        "TB": 1024 * 1024 * 1024 * 1024,
-        "K": 1024,
-        "M": 1024 * 1024,
-        "G": 1024 * 1024 * 1024,
-        "T": 1024 * 1024 * 1024 * 1024,
-    }
-
-    multiplier = multipliers.get(unit, 1024 * 1024)  # Default to MB
-    return int(value * multiplier)
-
-
-def validate_compression_settings(compression, compression_level, verbose=False):
-    """
-    Validate and normalize compression settings.
-
-    Args:
-        compression: Compression type string
-        compression_level: Compression level (can be None for defaults)
-        verbose: Whether to print verbose output
-
-    Returns:
-        tuple: (normalized_compression, validated_level, compression_desc)
-    """
-    compression = compression.upper()
-    valid_compressions = ["ZSTD", "GZIP", "BROTLI", "LZ4", "SNAPPY", "UNCOMPRESSED"]
-
-    if compression not in valid_compressions:
-        raise InvalidParameterError(
-            "compression",
-            f"Invalid compression '{compression}'. Must be one of: {', '.join(valid_compressions)}",
-        )
-
-    # Handle compression level based on format
-    compression_ranges = {
-        "GZIP": (1, 9, 6),  # min, max, default
-        "ZSTD": (1, 22, 15),  # min, max, default
-        "BROTLI": (1, 11, 6),  # min, max, default
-    }
-
-    if compression in compression_ranges:
-        min_level, max_level, default_level = compression_ranges[compression]
-
-        # Use default if not specified
-        if compression_level is None:
-            compression_level = default_level
-
-        if compression_level < min_level or compression_level > max_level:
-            raise InvalidParameterError(
-                "compression_level",
-                f"{compression} compression level must be between {min_level} and {max_level}, got {compression_level}",
-            )
-        compression_desc = f"{compression}:{compression_level}"
-    elif compression in ["LZ4", "SNAPPY"]:
-        if compression_level and compression_level != 15 and verbose:  # Not default
-            warn(
-                f"Note: {compression} does not support compression levels. Ignoring level {compression_level}."
-            )
-        compression_level = None  # These formats don't use compression levels
-        compression_desc = compression
-    else:
-        compression_level = None  # UNCOMPRESSED doesn't use levels
-        compression_desc = compression
-
-    return compression, compression_level, compression_desc
 
 
 def zm_suffix_sql(geom_expr: str, sep: str = " ") -> str:
@@ -576,328 +379,6 @@ def compute_geometry_dimensions_via_sql(
         if zm is not None and zm in _DUCKDB_ZMFLAG_TO_GEOARROW_DIM:
             dims.add(_DUCKDB_ZMFLAG_TO_GEOARROW_DIM[zm])
     return dims
-
-
-def _rebuild_array_with_type(
-    chunked_array,
-    new_type,
-):
-    """
-    Rebuild a chunked array with a new extension type.
-
-    This preserves CRS and other type metadata, unlike cast() which may reset them.
-    Used when applying CRS to geoarrow geometry arrays.
-
-    Args:
-        chunked_array: PyArrow ChunkedArray to rebuild
-        new_type: New PyArrow ExtensionType to apply
-
-    Returns:
-        pa.ChunkedArray: New chunked array with the new type
-    """
-    import pyarrow as pa
-
-    new_chunks = []
-    for chunk in chunked_array.chunks:
-        new_chunk = pa.ExtensionArray.from_storage(new_type, chunk.storage)
-        new_chunks.append(new_chunk)
-
-    return pa.chunked_array(new_chunks, type=new_type)
-
-
-def _type_category(arrow_type) -> str:
-    """Classify a PyArrow type into a single promotion category.
-
-    Categories are mutually exclusive, so two types in the same category are
-    treated alike by the promotion rules. Returns "other" for types that have
-    no special promotion handling (caller falls back to the first type).
-    """
-    import pyarrow as pa
-
-    if pa.types.is_null(arrow_type):
-        return "null"
-    if pa.types.is_boolean(arrow_type):
-        return "bool"
-    if pa.types.is_integer(arrow_type):
-        return "int"
-    if pa.types.is_floating(arrow_type):
-        return "float"
-    if pa.types.is_decimal(arrow_type):
-        return "decimal"
-    if pa.types.is_string(arrow_type) or pa.types.is_large_string(arrow_type):
-        return "string"
-    if pa.types.is_binary(arrow_type) or pa.types.is_large_binary(arrow_type):
-        return "binary"
-    if pa.types.is_timestamp(arrow_type):
-        return "timestamp"
-    return "other"
-
-
-def _promote_special(type_a, type_b, cat_a: str, cat_b: str):
-    """Promote null / binary / string categories. Returns None if neither applies.
-
-    Order matters: binary is handled before string so that binary geometry
-    data is never coerced to string.
-    """
-    import pyarrow as pa
-
-    if cat_a == "null":
-        return type_b
-    if cat_b == "null":
-        return type_a
-
-    # Binary types: don't promote to string (would corrupt geometry data)
-    if cat_a == "binary" or cat_b == "binary":
-        if cat_a == "binary" and cat_b == "binary":
-            return pa.large_binary()  # Safest binary type
-        return type_a  # Can't promote binary + non-binary safely
-
-    # String wins over numeric (safest fallback for text data)
-    if cat_a == "string" or cat_b == "string":
-        return pa.string()
-
-    return None
-
-
-def _promote_int_pair(type_a, type_b):
-    """Promote two integer types, guarding against uint64/signed overflow."""
-    import pyarrow as pa
-
-    a_unsigned = pa.types.is_unsigned_integer(type_a)
-    b_unsigned = pa.types.is_unsigned_integer(type_b)
-    # uint64 mixed with signed int → float64 (avoids overflow for large uint64)
-    if (a_unsigned and type_a == pa.uint64() and not b_unsigned) or (
-        b_unsigned and type_b == pa.uint64() and not a_unsigned
-    ):
-        return pa.float64()
-    return pa.int64()
-
-
-def _promote_numeric_pair(type_a, type_b, cat_a: str, cat_b: str):
-    """Promote bool / int / float / decimal combinations. None if not numeric."""
-    import pyarrow as pa
-
-    cats = {cat_a, cat_b}
-
-    # bool + int → int64 (bool can safely cast to int)
-    if cats == {"bool", "int"}:
-        return pa.int64()
-
-    # bool + float → float64
-    if cats == {"bool", "float"}:
-        return pa.float64()
-
-    # int + int → check for unsigned overflow risk
-    if cat_a == "int" and cat_b == "int":
-        return _promote_int_pair(type_a, type_b)
-
-    # Any remaining int/float/decimal mix → float64 (safest common type)
-    numeric = {"int", "float", "decimal"}
-    if cat_a in numeric and cat_b in numeric:
-        return pa.float64()
-
-    return None
-
-
-def _promote_timestamp_pair(type_a, type_b, cat_a: str, cat_b: str):
-    """Promote two timestamps to the finer unit. None if not both timestamps."""
-    if cat_a != "timestamp" or cat_b != "timestamp":
-        return None
-
-    # ns > us > ms > s — return the finer-grained unit (lower number = finer)
-    unit_order = {"ns": 0, "us": 1, "ms": 2, "s": 3}
-    if unit_order.get(type_a.unit, 99) <= unit_order.get(type_b.unit, 99):
-        return type_a
-    return type_b
-
-
-def _promote_numeric_type(type_a, type_b):
-    """
-    Compute a common type that both input types can safely cast to.
-
-    Type promotion rules (ordered by preference):
-    - Same types → return as-is
-    - null + any → any
-    - bool + int* → int64
-    - int8/16/32 + int64 → int64
-    - float32 + float64 → float64
-    - int* + float* → float64
-    - int* + decimal128 → float64
-    - decimal128 + float* → float64
-    - timestamp[*] + timestamp[*] → timestamp with finest unit
-    - any numeric + string → string (but not binary + string)
-    - incompatible types → first type (caller handles cast errors)
-
-    Args:
-        type_a: First PyArrow type
-        type_b: Second PyArrow type
-
-    Returns:
-        PyArrow type that both can safely cast to
-    """
-    if type_a == type_b:
-        return type_a
-
-    cat_a = _type_category(type_a)
-    cat_b = _type_category(type_b)
-
-    for handler in (_promote_special, _promote_numeric_pair, _promote_timestamp_pair):
-        result = handler(type_a, type_b, cat_a, cat_b)
-        if result is not None:
-            return result
-
-    # Fallback: return first type (caller handles cast errors with context)
-    return type_a
-
-
-def _compute_unified_schema(schemas: list):
-    """
-    Compute unified schema from multiple schemas with safe type promotion.
-
-    Used when merging paginated results where different pages may have
-    different inferred types for the same field (e.g., int64 vs decimal128).
-
-    Args:
-        schemas: List of PyArrow schemas to unify
-
-    Returns:
-        Unified PyArrow schema that all input schemas can safely cast to
-    """
-    import pyarrow as pa
-
-    if not schemas:
-        return pa.schema([])
-
-    if len(schemas) == 1:
-        return schemas[0]
-
-    # Build field name → list of types mapping
-    field_types: dict[str, list] = {}
-    field_order: list[str] = []
-
-    for schema in schemas:
-        for field in schema:
-            if field.name not in field_types:
-                field_types[field.name] = []
-                field_order.append(field.name)
-            field_types[field.name].append(field.type)
-
-    # Compute unified type for each field
-    unified_fields = []
-    for name in field_order:
-        types = field_types[name]
-        unified_type = types[0]
-        for t in types[1:]:
-            unified_type = _promote_numeric_type(unified_type, t)
-        unified_fields.append(pa.field(name, unified_type))
-
-    return pa.schema(unified_fields)
-
-
-def _parse_time_strings(col, target_type):
-    """Parse HH:MM:SS[.fff] strings into a time32/time64 array.
-
-    PyArrow has no string → time cast, but GDAL delivers ArcGIS TimeOnly
-    values as strings (ESRIJSON attributes, and all-null GeoJSON pages are
-    inferred as string), so the conversion must be explicit. Whole-second
-    values take the vectorized strptime path; fractional seconds fall back
-    to Python's ISO parser, which strptime cannot express.
-
-    Raises ValueError on a value that is not a time of day.
-    """
-    import datetime
-
-    import pyarrow as pa
-    import pyarrow.compute as pc
-
-    parsed = pc.strptime(col, format="%H:%M:%S", unit="ms", error_is_null=True)
-    unmatched = pc.and_(pc.is_valid(col), pc.is_null(parsed))
-    if pc.any(unmatched).as_py():
-        values = []
-        for value in col.to_pylist():
-            if value is None:
-                values.append(None)
-                continue
-            try:
-                values.append(datetime.time.fromisoformat(value))
-            except ValueError as e:
-                raise ValueError(f"{value!r} is not a time of day: {e}") from e
-        return pa.array(values, type=target_type)
-    return parsed.cast(target_type)
-
-
-def _cast_table_to_schema(table, target_schema, *, page_info: str | None = None):
-    """
-    Cast a table's columns to match target schema types.
-
-    Handles type conversions that PyArrow's concat_tables(promote=True) cannot,
-    such as int64 → float64 when another table has decimal128.
-
-    Args:
-        table: PyArrow table to cast
-        target_schema: Target schema with desired types
-        page_info: Optional context string for error messages (e.g., "page 3")
-
-    Returns:
-        Table with columns cast to target schema types
-
-    Raises:
-        ValueError: If a column cannot be cast to the target type
-    """
-    import pyarrow as pa
-
-    from geoparquet_io.core.logging_config import warn
-
-    if table.schema.equals(target_schema):
-        return table
-
-    new_columns = []
-    for field in target_schema:
-        if field.name in table.column_names:
-            col = table.column(field.name)
-            if col.type != field.type:
-                # Warn about precision loss for decimal → float64
-                if pa.types.is_decimal(col.type) and pa.types.is_floating(field.type):
-                    warn(
-                        f"Column '{field.name}' cast from {col.type} to {field.type} "
-                        f"may lose precision for large values"
-                    )
-                # PyArrow cannot cast int32 → timestamp directly (it needs int64 as
-                # an intermediate) and raises ArrowNotImplementedError. This happens
-                # when a source (e.g. DuckDB) infers an epoch-zero/null date column as
-                # int32 while the target schema declares timestamp. Upcast to int64
-                # first so the timestamp cast below succeeds (issue #516).
-                if (
-                    pa.types.is_timestamp(field.type)
-                    and pa.types.is_integer(col.type)
-                    and col.type.bit_width < 64
-                ):
-                    col = col.cast(pa.int64())
-                # Cast to target type with error context. pa.cast() raises siblings
-                # ArrowInvalid (bad value) and ArrowNotImplementedError (unsupported
-                # conversion); catch both so neither escapes uncontextualized.
-                # ValueError also covers the explicit time-string parse, which
-                # PyArrow cannot do as a cast.
-                try:
-                    if pa.types.is_time(field.type) and (
-                        pa.types.is_string(col.type) or pa.types.is_large_string(col.type)
-                    ):
-                        col = _parse_time_strings(col, field.type)
-                    else:
-                        col = col.cast(field.type, safe=True)
-                except (pa.ArrowInvalid, pa.ArrowNotImplementedError, ValueError) as e:
-                    context = f" ({page_info})" if page_info else ""
-                    raise ValueError(
-                        f"Failed to cast column '{field.name}' from {col.type} to "
-                        f"{field.type}{context}: {e}"
-                    ) from e
-            new_columns.append(col)
-        else:
-            # Missing column - create null array
-            null_array = pa.nulls(table.num_rows, type=field.type)
-            new_columns.append(null_array)
-
-    return pa.table(dict(zip([f.name for f in target_schema], new_columns, strict=True)))
 
 
 def _resolve_auto_version(detected: str | None) -> str | None:
@@ -3352,352 +2833,6 @@ def write_geoparquet_table(
             )
 
 
-def format_size(size_bytes):
-    """Convert bytes to human readable string."""
-    for unit in ["B", "KB", "MB", "GB"]:
-        if size_bytes < 1024:
-            return f"{size_bytes:.2f} {unit}"
-        size_bytes /= 1024
-    return f"{size_bytes:.2f} TB"
-
-
-#: Struct fields a bbox covering column must expose.
-_BBOX_REQUIRED_FIELDS = frozenset({"xmin", "ymin", "xmax", "ymax"})
-
-
-def _bbox_column_from_covering(geo_meta) -> str | None:
-    """Return the bbox column name referenced by GeoParquet ``covering.bbox``.
-
-    The covering metadata is the authoritative pointer to a file's bbox column
-    (its name need not follow any convention). Returns ``None`` when absent or
-    malformed.
-    """
-    if not isinstance(geo_meta, dict):
-        return None
-    columns = geo_meta.get("columns", {})
-    if not isinstance(columns, dict):
-        return None
-    for col_info in columns.values():
-        if not isinstance(col_info, dict):
-            continue
-        covering = col_info.get("covering")
-        bbox_refs = covering.get("bbox") if isinstance(covering, dict) else None
-        if (
-            isinstance(bbox_refs, dict)
-            and _BBOX_REQUIRED_FIELDS.issubset(bbox_refs)
-            and all(isinstance(ref, list) and len(ref) == 2 for ref in bbox_refs.values())
-        ):
-            return bbox_refs["xmin"][0]
-    return None
-
-
-def _schema_struct_child_names(schema_info, column_name) -> set | None:
-    """Child field names of struct column ``column_name`` from a flat
-    ``parquet_schema()`` listing; ``None`` if the column is absent or not a struct."""
-    for i, col in enumerate(schema_info):
-        if col.get("name") != column_name:
-            continue
-        num_children = col.get("num_children") or 0
-        if num_children < 1:
-            return None
-        return {
-            schema_info[i + j].get("name", "")
-            for j in range(1, num_children + 1)
-            if i + j < len(schema_info)
-        }
-    return None
-
-
-def _find_bbox_column_in_schema(schema_info, verbose):
-    """Find bbox column in schema by conventional names or structure.
-
-    Args:
-        schema_info: List of column dicts from get_schema_info()
-        verbose: Whether to print verbose output
-
-    Note:
-        DuckDB's parquet_schema() returns nested struct fields without parent prefix.
-        For a struct column 'bbox' with fields xmin/ymin/xmax/ymax:
-        - bbox appears with num_children=4
-        - Child fields appear as 'xmin', 'ymin', 'xmax', 'ymax' (not 'bbox.xmin')
-    """
-    # Check for columns ending with these suffixes (e.g., geometry_bbox, bbox)
-    conventional_suffixes = ["bbox", "bounds", "extent"]
-    required_fields = {"xmin", "ymin", "xmax", "ymax"}
-
-    for i, col in enumerate(schema_info):
-        name = col.get("name", "")
-        num_children = col.get("num_children", 0)
-
-        if not name:
-            continue
-
-        # Check if column name ends with conventional suffixes and has struct children
-        is_bbox_name = any(name.endswith(suffix) for suffix in conventional_suffixes)
-        if is_bbox_name and num_children >= 4:
-            # Get the next num_children entries as the struct's child fields
-            child_names = set()
-            for j in range(1, num_children + 1):
-                if i + j < len(schema_info):
-                    child_name = schema_info[i + j].get("name", "")
-                    child_names.add(child_name)
-
-            # Check if all required fields are present
-            if required_fields.issubset(child_names):
-                if verbose:
-                    debug(f"Found bbox column: {name} with children: {child_names}")
-                return name
-
-    return None
-
-
-def _check_bbox_metadata_covering(geo_meta, has_bbox_column, verbose, bbox_column_name=None):
-    """Check if geo metadata contains proper bbox covering.
-
-    Args:
-        geo_meta: Parsed geo metadata dict (from get_geo_metadata())
-        has_bbox_column: Whether a bbox column was found in schema
-        verbose: Whether to print verbose output
-        bbox_column_name: The bbox column actually present in the file. A
-            covering only counts as declaring *this* file's bbox column when it
-            names it; a covering pointing at some other (or absent) column is a
-            dangling reference, and treating it as "declared" let a broken file
-            pass `check` while the success message named a different column
-            entirely (#738).
-    """
-    if not (geo_meta and has_bbox_column):
-        return False
-
-    if verbose:
-        debug("\nParsed geo metadata:")
-        debug(json.dumps(geo_meta, indent=2))
-
-    # Validation-shared, so the block stays as the file really holds it (#883's
-    # line) and this guards instead of sanitizing: `gpio check bbox` and `check
-    # all` report through here. A `columns` that is not an object, or a
-    # `covering` that is not one, simply declares no covering -- the truthful
-    # answer, where iterating it raised an `AttributeError` from someone else's
-    # file (#947).
-    columns = geo_meta.get("columns") if isinstance(geo_meta, dict) else None
-    if isinstance(columns, dict):
-        for _col_name, col_info in columns.items():
-            covering = col_info.get("covering") if isinstance(col_info, dict) else None
-            if isinstance(covering, dict) and covering.get("bbox"):
-                bbox_refs = covering["bbox"]
-                # Check if the bbox covering has the required structure
-                if (
-                    isinstance(bbox_refs, dict)
-                    and all(key in bbox_refs for key in ["xmin", "ymin", "xmax", "ymax"])
-                    and all(isinstance(ref, list) and len(ref) == 2 for ref in bbox_refs.values())
-                ):
-                    referenced_bbox_column = bbox_refs["xmin"][0]
-                    if bbox_column_name is not None and referenced_bbox_column != bbox_column_name:
-                        if verbose:
-                            debug(
-                                f"Covering references column '{referenced_bbox_column}', "
-                                f"but the file's bbox column is '{bbox_column_name}' - "
-                                "treating the bbox column as undeclared"
-                            )
-                        continue
-                    if verbose:
-                        debug(
-                            f"Found bbox covering in metadata referencing column: {referenced_bbox_column}"
-                        )
-                    return True
-
-    return False
-
-
-def _determine_bbox_status(has_bbox_column, bbox_column_name, has_bbox_metadata):
-    """Determine bbox status and message."""
-    if has_bbox_column and has_bbox_metadata:
-        return "optimal", f"✓ Found bbox column '{bbox_column_name}' with proper metadata covering"
-    elif has_bbox_column:
-        return (
-            "suboptimal",
-            f"⚠️  Found bbox column '{bbox_column_name}' but no bbox covering metadata (recommended for better performance)",
-        )
-    else:
-        return "poor", "❌ No valid bbox column found"
-
-
-class BboxInfo(TypedDict, total=False):
-    """Bbox structure information returned by check_bbox_structure."""
-
-    has_bbox_column: bool
-    bbox_column_name: str | None
-    has_bbox_metadata: bool
-    status: Literal["optimal", "suboptimal", "poor", "native"]
-    message: str
-
-
-def check_bbox_structure(parquet_file, verbose=False) -> BboxInfo:
-    """
-    Check bbox structure and metadata coverage in a GeoParquet file.
-
-    Returns:
-        dict: Results including:
-            - has_bbox_column (bool): Whether a valid bbox struct column exists
-            - bbox_column_name (str): Name of the bbox column if found
-            - has_bbox_metadata (bool): Whether bbox covering is specified in metadata
-            - status (str): "optimal", "suboptimal", or "poor"
-            - message (str): Human readable description
-    """
-    from geoparquet_io.core.duckdb_metadata import get_geo_metadata, get_schema_info
-
-    # Get schema info using DuckDB
-    schema_info = get_schema_info(parquet_file)
-
-    if verbose:
-        debug("\nSchema fields:")
-        for col in schema_info:
-            name = col.get("name", "")
-            col_type = col.get("type", "")
-            if name:  # Skip empty names
-                debug(f"  {name}: {col_type}")
-
-    # Find the bbox column: the authoritative covering metadata first (spec-valid
-    # files may use non-conventional names), then the naming-convention fallback.
-    geo_meta = get_geo_metadata(parquet_file)
-    bbox_column_name = None
-    covering_column = _bbox_column_from_covering(geo_meta)
-    if covering_column:
-        children = _schema_struct_child_names(schema_info, covering_column)
-        if children and _BBOX_REQUIRED_FIELDS.issubset(children):
-            bbox_column_name = covering_column
-            if verbose:
-                debug(f"Found bbox column from covering metadata: {covering_column}")
-    if bbox_column_name is None:
-        bbox_column_name = _find_bbox_column_in_schema(schema_info, verbose)
-    has_bbox_column = bbox_column_name is not None
-
-    # Check for bbox covering in the geo metadata
-    has_bbox_metadata = _check_bbox_metadata_covering(
-        geo_meta, has_bbox_column, verbose, bbox_column_name
-    )
-
-    # Determine status and message
-    status, message = _determine_bbox_status(has_bbox_column, bbox_column_name, has_bbox_metadata)
-
-    if verbose:
-        debug("\nFinal results:")
-        debug(f"  has_bbox_column: {has_bbox_column}")
-        debug(f"  bbox_column_name: {bbox_column_name}")
-        debug(f"  has_bbox_metadata: {has_bbox_metadata}")
-        debug(f"  status: {status}")
-        debug(f"  message: {message}")
-
-    return {
-        "has_bbox_column": has_bbox_column,
-        "bbox_column_name": bbox_column_name if has_bbox_column else None,
-        "has_bbox_metadata": has_bbox_metadata,
-        "status": status,
-        "message": message,
-    }
-
-
-def get_bbox_advice(
-    parquet_file: str,
-    operation: str,
-    verbose: bool = False,
-) -> dict:
-    """
-    Get version-aware bbox optimization advice.
-
-    Provides context-aware recommendations based on file type and operation:
-    - For GeoParquet 2.0/parquet-geo with spatial_filtering: No bbox pre-filter
-      needed -- a bare ST_Intersects ON clause lets DuckDB's SPATIAL_JOIN operator
-      engage, which is dramatically faster than a bbox-overlap NL join (issue #538).
-    - For GeoParquet 2.0/parquet-geo with bounds_calculation: bbox still recommended (faster)
-    - For GeoParquet 1.x without bbox: Suggest adding bbox OR upgrading to 2.0
-
-    Args:
-        parquet_file: Path to the parquet file
-        operation: One of:
-            - "spatial_filtering": For ST_Intersects, spatial joins, etc.
-            - "bounds_calculation": For centroid, extent, quadkey, etc.
-            - "check": For validation/inspection
-        verbose: Whether to print verbose output
-
-    Returns:
-        dict with:
-            - needs_warning: bool - Whether to show a warning to the user
-            - skip_bbox_prefilter: bool - Whether to skip bbox pre-filtering in queries.
-              Only True for spatial_filtering with native geometry. There a bare
-              ST_Intersects ON clause is recognized by DuckDB's SPATIAL_JOIN operator
-              and is much faster than ANDing a bbox-overlap test in front of it, which
-              defeats SPATIAL_JOIN and forces a BLOCKWISE_NL_JOIN (issue #538). For
-              bounds_calculation, always False since the bbox column provides
-              pre-computed values that are faster than geometry stats.
-            - has_native_geometry: bool - Whether file uses native Parquet geometry types
-            - message: str - User-facing message (if needs_warning)
-            - suggestions: list[str] - Suggested actions for the user
-    """
-    file_info = detect_geoparquet_file_type(parquet_file, verbose)
-    bbox_info = check_bbox_structure(parquet_file, verbose)
-
-    # Read native-geometry capability directly rather than inferring from file_type:
-    # a 1.1-geoarrow file carries native geometry columns but is labelled
-    # "geoparquet_v1" (version starts with "1."), so a file_type membership check
-    # would misclassify it as non-native. .get() also avoids a KeyError.
-    has_native_geo = file_info.get("has_native_geo_types", False)
-    has_bbox = bbox_info["has_bbox_column"]
-
-    # Only skip bbox pre-filtering for spatial_filtering operations with native geometry.
-    # For bounds_calculation, bbox column provides pre-computed values that are faster.
-    skip_bbox = has_native_geo and operation == "spatial_filtering"
-
-    result = {
-        "needs_warning": False,
-        "skip_bbox_prefilter": skip_bbox,
-        "has_native_geometry": has_native_geo,
-        "has_bbox_column": has_bbox,
-        "bbox_column_name": bbox_info.get("bbox_column_name"),
-        "message": "",
-        "suggestions": [],
-    }
-
-    if operation == "spatial_filtering":
-        if has_native_geo:
-            # Native geometry -> bare ST_Intersects ON clause engages DuckDB's
-            # SPATIAL_JOIN operator (the fast path, issue #538) - no warning needed.
-            if verbose:
-                debug("Native geometry: bare ST_Intersects engages DuckDB SPATIAL_JOIN")
-        elif not has_bbox:
-            # 1.x without bbox - warn and suggest options
-            result["needs_warning"] = True
-            result["message"] = "No bbox column found"
-            result["suggestions"] = [
-                "Add a bbox column: gpio add bbox <file>",
-                "Or upgrade to GeoParquet 2.0: gpio convert <file> --geoparquet-version 2.0",
-            ]
-
-    elif operation == "bounds_calculation":
-        # bbox column is still faster for bounds/centroid calculation (pre-computed values)
-        if not has_bbox:
-            result["needs_warning"] = True
-            result["message"] = "No bbox column - computing from geometry (slower)"
-            result["suggestions"] = [
-                "Add a bbox column for 3-4x faster bounds/centroid: gpio add bbox <file>"
-            ]
-
-    elif operation == "check":
-        if has_native_geo:
-            # Native geometry - bbox optional but can help with bounds queries
-            if not has_bbox and verbose:
-                debug("Native geometry type detected - bbox column optional for spatial queries")
-        elif not has_bbox:
-            # 1.x without bbox
-            result["needs_warning"] = True
-            result["message"] = "No bbox column found"
-            result["suggestions"] = [
-                "Add a bbox column: gpio add bbox <file>",
-                "Or upgrade to GeoParquet 2.0: gpio convert <file> --geoparquet-version 2.0",
-            ]
-
-    return result
-
-
 def _build_bounds_query(parquet_path, bbox_info, geometry_column, verbose):
     """Build query for bounds calculation.
 
@@ -4066,73 +3201,3 @@ def add_bbox(parquet_file, bbox_column_name="bbox", verbose=False):
         if os.path.exists(temp_file):
             os.remove(temp_file)
         raise GeoParquetError(f"Failed to add bbox: {str(e)}") from e
-
-
-def create_shapefile_zip(shapefile_path: str | Path, verbose: bool = False) -> Path:
-    """
-    Create a zip archive containing a shapefile and all its sidecar files.
-
-    Shapefiles consist of multiple files with different extensions (.shp, .shx, .dbf, .prj, .cpg).
-    This function creates a single .shp.zip archive containing all related files.
-
-    Args:
-        shapefile_path: Path to the main .shp file
-        verbose: Print verbose output
-
-    Returns:
-        Path to the created .shp.zip file
-
-    Raises:
-        click.ClickException: If shapefile or required sidecars are missing
-    """
-    import zipfile
-
-    configure_verbose(verbose)
-    shapefile_path = Path(shapefile_path)
-
-    if not shapefile_path.exists():
-        raise FileNotFoundGeoParquetError(str(shapefile_path))
-
-    # Shapefile extensions: .shp (main), .shx (index), .dbf (attributes) are required
-    # Optional: .prj (projection), .cpg (encoding), .sbn/.sbx (spatial index)
-    stem = shapefile_path.stem
-    parent = shapefile_path.parent
-
-    # Find all files with the same stem
-    sidecar_files = list(parent.glob(f"{stem}.*"))
-
-    # Filter to only shapefile-related extensions
-    shapefile_extensions = {".shp", ".shx", ".dbf", ".prj", ".cpg", ".sbn", ".sbx"}
-    sidecar_files = [f for f in sidecar_files if f.suffix.lower() in shapefile_extensions]
-
-    if not sidecar_files:
-        raise FileNotFoundGeoParquetError(str(shapefile_path), "no shapefile components found")
-
-    # Verify required files exist
-    required_extensions = {".shp", ".shx", ".dbf"}
-    found_extensions = {f.suffix.lower() for f in sidecar_files}
-    missing = required_extensions - found_extensions
-
-    if missing:
-        raise FileNotFoundGeoParquetError(
-            str(shapefile_path), f"missing required shapefile components: {', '.join(missing)}"
-        )
-
-    # Create zip file
-    zip_path = parent / f"{stem}.shp.zip"
-
-    if verbose:
-        debug(f"Creating shapefile archive: {zip_path}")
-        debug(f"Including {len(sidecar_files)} files: {[f.name for f in sidecar_files]}")
-
-    with zipfile.ZipFile(zip_path, "w", zipfile.ZIP_DEFLATED) as zipf:
-        for sidecar in sidecar_files:
-            zipf.write(sidecar, sidecar.name)
-            if verbose:
-                debug(f"  Added: {sidecar.name}")
-
-    if verbose:
-        zip_size = zip_path.stat().st_size
-        success(f"Created shapefile archive: {zip_path} ({format_size(zip_size)})")
-
-    return zip_path
